frontend/pyproject.toml
[project]
name = "desafio-rag-frontend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.35.0",
    "requests>=2.32.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.2",
]
